# api/openapi.yaml
openapi: 3.0.3
info:
  title: Ad Bidding Service API
  description: API for managing ad line items and serving winning ads based on various criteria
  version: 1.0.0
  contact:
    name: Your Company
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/v1/ads:
    get:
      summary: Get winning ads for a placement
      description: Returns the winning ads for a specific placement with optional filters (TO BE IMPLEMENTED BY CANDIDATE)
      operationId: getWinningAds
      parameters:
        - name: placement
          in: query
          description: Target placement identifier
          required: true
          schema:
            type: string
        - name: category
          in: query
          description: Filter by category
          required: false
          schema:
            type: string
        - name: keyword
          in: query
          description: Filter by keyword
          required: false
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of ads to return
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
            maximum: 10
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ad'
        400:
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Ad:
      type: object
      required:
        - id
        - name
        - advertiser_id
        - bid
      properties:
        id:
          type: string
          description: Line item ID
          example: "li_1234567890"
        name:
          type: string
          description: Display name of the ad
          example: "Summer Sale Banner"
        advertiser_id:
          type: string
          description: ID of the advertiser
          example: "adv123"
        bid:
          type: number
          format: float
          description: Actual bid amount for this impression
          example: 2.3
        placement:
          type: string
          description: Placement where the ad will be shown
          example: "homepage_top"
        serve_url:
          type: string
          description: URL to serve for this ad
          example: "/api/v1/lineitems/li_1234567890"
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: Error code
          example: 400
        message:
          type: string
          description: Error message
          example: "Invalid request parameters"
        details:
          type: object
          description: Additional error details
          example:
            field: "bid"
            reason: "must be a positive number"